Split resolver replay declaration tasks
